// web/src/api/generated/models/ExportStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ExportStatus = {
    QUEUED: 'QUEUED',
    RUNNING: 'RUNNING',
    READY: 'READY',
    FAILED: 'FAILED',
    EXPIRED: 'EXPIRED'
} as const;
export type ExportStatus = typeof ExportStatus[keyof typeof ExportStatus];


export function instanceOfExportStatus(value: any): boolean {
    for (const key in ExportStatus) {
        if (Object.prototype.hasOwnProperty.call(ExportStatus, key)) {
            if (ExportStatus[key as keyof typeof ExportStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ExportStatusFromJSON(json: any): ExportStatus {
    return ExportStatusFromJSONTyped(json, false);
}

export function ExportStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExportStatus {
    return json as ExportStatus;
}

export function ExportStatusToJSON(value?: ExportStatus | null): any {
    return value as any;
}

export function ExportStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): ExportStatus {
    return value as ExportStatus;
}